fix: resolve PR review comments — remove handler wrappers and inline calls
- Inline parse+call in run.rs event_bridge_handler (single-line form per reviewer suggestion)
- Remove handlers.rs — merge handler logic directly into fan_out functions
- Refactor fan_out_timeline_entries, fan_out_popular_post, fan_out_popular_space to accept typed models (Post/SpaceCommon) directly
- Update call sites in update_post.rs and like_post.rs to pass model directly
