+++ app/ratel/src/features/posts/controllers/like_post.rs
@@ -22,22 +22,11 @@ pub async fn like_post_handler(post_id: FeedPartition, like: bool) -> Result<Lik
         {
             // Check if the post just became popular and trigger broader fan-out
             if let Some(post) = Post::get(cli, &post_pk, Some(EntityType::Post)).await? {
-                if crate::features::timeline::services::is_popular(
-                    post.likes,
-                    post.comments,
-                    post.shares,
-                ) {
-                    let _ = crate::features::timeline::services::fan_out_popular_post(
-                        cli,
-                        &post.pk,
-                        &post.user_pk,
-                        post.updated_at,
-                    )
+                let _ = crate::features::timeline::services::fan_out_popular_post(post)
                     .await
                     .map_err(|e| {
                         tracing::error!("popular post fan-out failed: {}", e);
                     });
-                }
             }
         }
