+++ app/ratel/src/features/timeline/services/mod.rs
@@ -1,5 +1,3 @@
 mod fan_out;
-mod handlers;
 
 pub use fan_out::*;
-pub use handlers::*;
